Refactor decision maker module, separating interface and implementation
