+++ aea/decision_maker/active.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2018-2019 Fetch.AI Limited
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""This module contains the pointers to the active decision maker class and auxilliary classes."""
+
+from aea.decision_maker.default import DecisionMaker  # noqa
+from aea.decision_maker.default import GoalPursuitReadiness  # noqa
+from aea.decision_maker.default import LedgerStateProxy  # noqa
+from aea.decision_maker.default import OwnershipState  # noqa
+from aea.decision_maker.default import Preferences  # noqa
