+++ packages/hashes.csv
@@ -56,7 +56,7 @@ fetchai/skills/scaffold,QmdHM1aaUSajF5C375wtjt3yLFpXjmDYKawLkAs9KkwrYu
 fetchai/skills/simple_service_registration,QmY3x3BbGEFCtFn8W4ZTmZCHYsVTUTZESXgUa6ftMhuPdN
 fetchai/skills/tac_control,QmPEv7meAPsJTq6f4vZrviXqMzjppMYt26sC6Z8wNVQkFa
 fetchai/skills/tac_control_contract,QmXYgqrTSb7vp2xpmuECkzaCQrQUGCgpmutUvsWZXfrhr1
-fetchai/skills/tac_negotiation,QmagQNeBtnhaFVgMn5bqdLrZGmGKDFAwhggSXRzzA6idKn
+fetchai/skills/tac_negotiation,Qma8nhzoKro9YXigt7aC3zAYgbRCYwcmmVNBxZ6BdZsSzb
 fetchai/skills/tac_participation,Qme2Pc9Vb3ePjerrrEz6eArSRub2pYNtPP5GpnyhDHptai
 fetchai/skills/thermometer,Qmc95kXNiDoyQXXRGDd2eXPQhP4EQX7poCuk3V3zWHUyFg
 fetchai/skills/thermometer_client,QmUtcQYnu3fZiywbeFFYqbYczxXhfPQrauF8ru2EhZxbzo
